refactor: extract country checker interface (health-declaration-step)
- Move HealthDeclarationProperties to http/config/feature and extract interface to service/checkin/feature
- Add country checker mock for tests
