+++ http/src/main/java/com/aterrizar/http/config/feature/HealthDeclarationProperties.java
@@ -1,27 +1,22 @@
-package com.aterrizar.service.checkin.config;
+package com.aterrizar.http.config.feature;
 
 import java.util.List;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
+import com.aterrizar.service.checkin.feature.HealthDeclarationCountryChecker;
 import com.neovisionaries.i18n.CountryCode;
 
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Configuration
 @Data
 @ConfigurationProperties(prefix = "feature.health.declaration.enabled")
-@NoArgsConstructor
-public class HealthDeclarationProperties {
+public class HealthDeclarationProperties implements HealthDeclarationCountryChecker {
     private List<CountryCode> countries;
 
-    public HealthDeclarationProperties(CountryCode... countries) {
-        this(List.of(countries));
-    }
-
-    public HealthDeclarationProperties(List<CountryCode> countries) {
-        this.countries = countries;
+    public boolean isCountryEnabled(CountryCode countryCode) {
+        return this.countries.contains(countryCode);
     }
 }
